add pad prerocess exec
