+++ src/plugins/intel_cpu/src/nodes/executors/common/ref_interpolate.cpp
@@ -11,7 +11,7 @@ void ov::intel_cpu::RefInterpolateExecutor::exec(const std::vector<MemoryCPtr>& 
     size_t N = srcDimPad5d[0], C = srcDimPad5d[1], ID = srcDimPad5d[2], IH = srcDimPad5d[3], IW = srcDimPad5d[4];
     size_t OD = dstDim5d[2], OH = dstDim5d[3], OW = dstDim5d[4];
 
-    auto in_ptr_ = static_cast<const uint8_t *>(src[0]->GetPtr());
+    auto in_ptr_ = padPreprocess(src, dst);
     auto out_ptr_ = static_cast<uint8_t *>(dst[0]->GetPtr());
 
     switch (interpAttrs.mode) {
