+++ src/plugins/intel_cpu/src/nodes/executors/x64/jit_interpolate.cpp
@@ -1664,7 +1664,7 @@ bool ov::intel_cpu::JITInterpolateExecutor::init(const InterpolateAttrs &interpo
 }
 
 void ov::intel_cpu::JITInterpolateExecutor::exec(const std::vector<MemoryCPtr>& src, const std::vector<MemoryPtr>& dst, const void *post_ops_data_) {
-    auto in_ptr_ = static_cast<const uint8_t *>(src[0]->GetPtr());
+    auto in_ptr_ = padPreprocess(src, dst);
     auto out_ptr_ = static_cast<uint8_t *>(dst[0]->GetPtr());
     size_t N = srcDimPad5d[0], C = srcDimPad5d[1], ID = srcDimPad5d[2], IH = srcDimPad5d[3], IW = srcDimPad5d[4];
     size_t OD = dstDim5d[2], OH = dstDim5d[3], OW = dstDim5d[4];
